FossilOrigin-Name: 85ebd46c701e0a094a4690cd8f1d0cbae9aa257c

+++ ext/fts5/fts5_expr.c
@@ -430,7 +430,7 @@ static int fts5ExprPhraseIsMatch(
     }
     if( rc!=SQLITE_OK ) goto ismatch_out;
     sqlite3Fts5PoslistReaderInit(a, n, &aIter[i]);
-    aIter[i].bFlag = bFlag;
+    aIter[i].bFlag = (u8)bFlag;
     if( aIter[i].bEof ) goto ismatch_out;
   }
